Convert "common" directory into a dynamic library
Part of #1218
