+++ libs/CCAppCommon/src/CMakeLists.txt
@@ -1,0 +1,12 @@
+
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/ccApplicationBase.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccCameraParamEditDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccDisplayOptionsDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccOptions.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccPickOneElementDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccPluginManager.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccStereoModeDlg.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccTranslationManager.cpp
+)
