+++ libs/CCAppCommon/src/ccApplicationBase.cpp
