+++ libs/CCAppCommon/src/ccDisplayOptionsDlg.cpp
