+++ libs/CCAppCommon/src/ccPluginManager.cpp
