+++ libs/CCAppCommon/src/ccStereoModeDlg.cpp
