+++ libs/CCAppCommon/src/ccTranslationManager.cpp
