+++ libs/CCAppCommon/ui/CMakeLists.txt
@@ -1,0 +1,12 @@
+
+target_sources( ${PROJECT_NAME}
+	PRIVATE
+		${CMAKE_CURRENT_LIST_DIR}/cameraParamDlg.ui
+		${CMAKE_CURRENT_LIST_DIR}/displayOptionsDlg.ui
+		${CMAKE_CURRENT_LIST_DIR}/pickOneElementDlg.ui
+		${CMAKE_CURRENT_LIST_DIR}/stereoModeDlg.ui
+)
+
+set_target_properties( ${PROJECT_NAME} PROPERTIES
+	AUTOUIC_SEARCH_PATHS ${CMAKE_CURRENT_LIST_DIR}
+)
