+++ libs/CCAppCommon/ui/cameraParamDlg.ui
