+++ libs/CCAppCommon/ui/displayOptionsDlg.ui
