+++ libs/CCAppCommon/ui/stereoModeDlg.ui
