+++ qCC/CMakeLists.txt
@@ -4,7 +4,6 @@ include( CMakePolicies NO_POLICY_SCOPE )
 
 include_directories( ${CloudComparePlugins_SOURCE_DIR} )
 include_directories( ${CMAKE_CURRENT_SOURCE_DIR} )
-include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/../common )
 include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/db_tree )
 include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/ui_templates )
 include_directories( ${CMAKE_CURRENT_BINARY_DIR} )
@@ -17,7 +16,6 @@ file( GLOB qrc_list *.qrc )
 
 file( GLOB txt_list TODO.txt ${CloudCompareProjects_SOURCE_DIR}/CHANGELOG.md )
 
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../common ${CMAKE_CURRENT_BINARY_DIR}/common)
 add_subdirectory( db_tree )
 add_subdirectory( pluginManager )
 
@@ -49,8 +47,7 @@ else()
 endif()
 
 target_link_libraries( ${PROJECT_NAME}
-    QCC_GL_LIB
-    CCPluginAPI
+    CCAppCommon
     QCustomPlot
     Qt5::OpenGL
     Qt5::PrintSupport
